vendorflow/requirements.txt
httpx>=0.27.0
pydantic>=2.0.0
PyMuPDF>=1.24.0
python-dotenv>=1.0.0
streamlit>=1.35.0
fpdf2>=2.7.0
PyYAML>=6.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
Pillow>=10.0.0
aiofiles>=23.0.0
email-validator>=2.0.0
nest_asyncio>=1.6.0